+++ packages/client/ui/react-ui/src/hooks/useAuth.ts
@@ -1,8 +1,8 @@
 import { useContext } from "react";
 
-import { AuthContext, type AuthContextType } from "../providers";
+import { AuthContext } from "../providers";
 
-export function useAuth(): AuthContextType {
+export function useAuth() {
     const context = useContext(AuthContext);
     if (context === undefined) {
         throw new Error("useAuth must be used within an AuthProvider");