+++ packages/client/ui/react-ui/src/providers/auth/web3/DynamicWeb3WalletConnect.tsx
@@ -1,7 +1,7 @@
 import type React from "react";
+import type { APIKeyEnvironmentPrefix } from "@crossmint/common-sdk-base";
 import { EthereumWalletConnectors } from "@dynamic-labs/ethereum";
 import { SolanaWalletConnectors } from "@dynamic-labs/solana";
-import type { APIKeyEnvironmentPrefix } from "@crossmint/common-sdk-base";
 import DynamicContextProviderWrapper from "@/components/dynamic-xyz/DynamicContextProviderWrapper";
 import { useCrossmintAuth } from "@/hooks/useCrossmintAuth";
 import { useAuthForm } from "../AuthFormProvider";
@@ -20,8 +20,8 @@ export function DynamicWeb3WalletConnect({
         <DynamicContextProviderWrapper
             apiKeyEnvironment={apiKeyEnvironment}
             settings={{
-                cssOverrides,
                 walletConnectors: [EthereumWalletConnectors, SolanaWalletConnectors],
+                cssOverrides,
                 events: {
                     onAuthFlowCancel() {
                         console.log("[CryptoWalletConnectionHandler] onAuthFlowCancel");
@@ -51,7 +51,7 @@ export function DynamicWeb3WalletConnect({
 
                         try {
                             const res = await crossmintAuth?.signInWithSmartWallet(address);
-                            const signature = (await wallet.connector?.signMessage(res.challenge)) as `0x${string}`;
+                            const signature = (await wallet.signMessage?.(res.challenge)) as `0x${string}`;
                             const authResponse = (await crossmintAuth?.authenticateSmartWallet(address, signature)) as {
                                 oneTimeSecret: string;
                             };