+++ apps/wallets/smart-wallet/next/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/client-sdk-smart-wallet-next-starter
 
+## 0.1.77
+
+### Patch Changes
+
+- 063ae65: Upgrade dynamic to v4
+- Updated dependencies [063ae65]
+  - @crossmint/client-sdk-react-ui@1.15.3
+
 ## 0.1.76
 
 ### Patch Changes